(b) hickey + lowy: one reading of a row's line, one owner of the paired obligation
Two findings, and they are the same shape from two sides — a whole that had
been split, and a composition two callers were writing by hand.

**Fragmentation · duplicated derivation — "where the lines are" was written
twice.** Both gestures begin by asking the page the same question: the same
`querySelectorAll`, the same `getBoundingClientRect`, the same `+ scrollX/Y`.
The drag owned the attribute name and the sweep imported `ROW_KEY` from it,
which SAID that measuring a row belongs to dragging one. It does not — it
belongs to the ROW. `drag/lines.ts` is that reading, with the two facts about it
that are easy to get wrong stated once (the mark is on the LINE and not on the
`<li>`, whose box is a subtree's; the coordinates are the DOCUMENT's, so the
answer survives the page moving under a live gesture).

Lowy's naming test is what settles the placement: the axis is "how a drawn row
is found and read", which changes when the row markup or the coordinate space
does — not when a drop or a pick changes. And the collapse goes one level
deeper than the walk: `Placed` now EXTENDS `Line` rather than repeating five of
its fields, so the drop planner's extra facts ride ON the measurement instead of
beside it. Two flat records sharing five fields is one concept written twice.

**Lowy §6.5 · the missing primitive is the composition.** `autoscroll.ts` and
`pointer.ts` are two real axes, and both new gestures were hand-composing them
identically — feed the scroller on every move, and stop it on every way the
gesture can end. That second half is a PAIRED OBLIGATION, which is precisely
what `pointer.ts` already took over from its callers for the threshold and the
teardown ("answered here so that a caller cannot answer it differently"). So it
is `Gesture.onPage` now: ask for the pointer in document coordinates and the
page keeps up, with the frame loop dying with the gesture rather than with
whoever remembered. Opt-in, because the third consumer must NOT have it — a
panel edge scrolling the outline behind it would be this option used by
accident — and `onMove` became optional, since the two page-aimed gestures no
longer want the window's coordinates at all.

Both fixes shrink the callers: `dragging.ts` loses the box arithmetic and the
scroller wiring, `sweeping.ts` loses a whole `measure` and its import of the
drag. `packages/web/README.md`'s walk-through gains the two modules.
